Finalizado Curso: Spring Boot API REST: Construa uma API

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -32,7 +33,7 @@ public class TopicosController {
 
 	@Autowired
 	private TopicoRepository topicoRepository;
-	
+
 	@Autowired
 	private CursoRepository cursoRepository;
 
@@ -54,59 +55,58 @@ public class TopicosController {
 	public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form, UriComponentsBuilder uriBuilder) {
 		Topico topico = form.converter(cursoRepository);
 		topicoRepository.save(topico);
-		
+
 		URI uri = uriBuilder.path("/topico/{id}").buildAndExpand(topico.getId()).toUri();
 		return ResponseEntity.created(uri).body(new TopicoDto(topico));
 	}
-	
-	
 
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
+	public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
 		
-		Topico topico = topicoRepository.getOne(id);
+		Optional<Topico> topico = topicoRepository.findById(id);
 		
-		return new DetalhesDoTopicoDto(topico);
+		if(topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));			
+		}else {
+			return ResponseEntity.notFound().build();
+		}
+		
 		
 	}
-	
-	
-	
+
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-		
-		Topico topico = form.atualizar(id, topicoRepository);
-		
-		
-		return ResponseEntity.ok(new TopicoDto(topico));
+
+		Optional<Topico> optional = topicoRepository.findById(id);
+
+		if (optional.isPresent()) {
+
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+
+		} else {
+			return ResponseEntity.notFound().build();
+		}
+
 	}
-	
-	
+
 	@DeleteMapping("/{id}")
 	@Transactional
-	public ResponseEntity<?> remover(@PathVariable Long id){
+	public ResponseEntity<?> remover(@PathVariable Long id) {
+
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
 		
 		topicoRepository.deleteById(id);
-		
-		
+
 		return ResponseEntity.ok().build();
+		}else {
+			return ResponseEntity.notFound().build();
+		}
+		
+		
+
 	}
-	
 
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
